requirements.txt
langchain>=0.2.16
langchain-community>=0.2.16
chromadb>=0.5.5,<1.0.0
sentence-transformers>=3.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.2
openai>=1.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-experimental>=0.0.61
pypdf>=4.0.0
